# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects/{subjectId}/chapters:
    get:
      operationId: getChapters
      tags: [chapters]
      summary: List chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{chapterId}:
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter (admin only)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter (admin only)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Topics ────────────────────────────────────────────────────────────
...
  /exam/start:
    post:
      operationId: startExam
      tags: [exams]
      summary: Start an exam attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamStartInput"
      responses:
        "201":
          description: Exam started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
        "403":
          description: Prerequisite gate not met
  /exam/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSubmitInput"
      responses:
        "200":
          description: Submitted and scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /exam/results/{resultId}:
    get:
      operationId: getExamResult
      tags: [exams]
      summary: Fetch exam result by ID
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
...
  /quizzes/{quizId}/questions:
    get:
      operationId: getQuizQuestions
      tags: [exams]
      summary: Get questions for a quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /quizzes:
    post:
      operationId: createQuiz
      tags: [exams]
      summary: Create a quiz (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /questions:
    post:
      operationId: createQuestion
      tags: [exams]
      summary: Create a question (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/bulk-import:
    post:
      operationId: bulkImportQuestions
      tags: [exams]
      summary: Bulk import questions from JSON array (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkQuestionsInput"
      responses:
        "201":
          description: Questions imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"

  # ── QR Code ───────────────────────────────────────────────────────────
  /qr/generate:
    post:
      operationId: generateQr
      tags: [exams]
      summary: Generate QR code for a YouTube URL (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrInput"
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrResult"

  # ── Notes / PDF ───────────────────────────────────────────────────────
...
  /admin/users/{userId}/approve:
    post:
      operationId: adminApproveUser
      tags: [admin]
      summary: Approve a pending user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/reject:
    post:
      operationId: adminRejectUser
      tags: [admin]
      summary: Reject / ban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/reset-progress:
    post:
      operationId: adminResetUserProgress
      tags: [admin]
      summary: Reset a user's progress (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetProgressInput"
      responses:
        "200":
          description: Progress reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Chapter:
      type: object
      properties:
        id:
          type: string
        subject_id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        is_active:
          type: boolean
      required: [id, subject_id, title, order_index, is_active]
    ChapterInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        is_active:
          type: boolean
      required: [title, order_index]
    ChapterUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        is_active:
          type: boolean
...
    ExamStartInput:
      type: object
      properties:
        quiz_id:
          type: string
      required: [quiz_id]
    ExamSession:
      type: object
      properties:
        attempt_id:
          type: string
        quiz_id:
          type: string
        started_at:
          type: string
        duration_minutes:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [attempt_id, quiz_id, started_at, duration_minutes, questions]
    ExamSubmitInput:
      type: object
      properties:
        attempt_id:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required: [attempt_id, answers]
    AnswerInput:
      type: object
      properties:
        question_id:
          type: string
        selected_option:
          type: ["string", "null"]
        time_spent_ms:
          type: integer
        is_marked_for_review:
          type: boolean
      required: [question_id]
    ExamResult:
      type: object
      properties:
        attempt_id:
          type: string
        quiz_id:
          type: string
        score:
          type: number
        total_marks:
          type: number
        accuracy:
          type: number
        time_taken_ms:
          type: integer
        negative_marks_applied:
          type: number
        passed:
          type: boolean
        status:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerResult"
        submitted_at:
          type: string
      required: [attempt_id, quiz_id, score, total_marks, accuracy, time_taken_ms, passed, status, answers, submitted_at]
    AnswerResult:
      type: object
      properties:
        question_id:
          type: string
        selected_option:
          type: ["string", "null"]
        correct_answer:
          type: string
        is_correct:
          type: boolean
        time_spent_ms:
          type: integer
        explanation:
          type: ["string", "null"]
        video_solution_url:
          type: ["string", "null"]
        qr_code_url:
          type: ["string", "null"]
      required: [question_id, correct_answer, is_correct, time_spent_ms]
...
    Question:
      type: object
      properties:
        id:
          type: string
        quiz_id:
          type: string
        question_text:
          type: string
        options:
          type: object
          additionalProperties:
            type: string
        correct_answer:
          type: string
        explanation:
          type: ["string", "null"]
        difficulty:
          type: integer
        order_index:
          type: integer
        video_solution_url:
          type: ["string", "null"]
        qr_code_url:
          type: ["string", "null"]
      required: [id, quiz_id, question_text, options, correct_answer, difficulty, order_index]
    Quiz:
      type: object
      properties:
        id:
          type: string
        topic_id:
          type: ["string", "null"]
        chapter_id:
          type: ["string", "null"]
        subject_id:
          type: ["string", "null"]
        title:
          type: string
        type:
          type: string
          enum: [lecture_quiz, dpp, pyqs, topic_test, chapter_test, subject_test, grand_test]
        passing_score:
          type: number
        max_attempts:
          type: integer
        duration_minutes:
          type: integer
        negative_marking:
          type: number
        is_active:
          type: boolean
      required: [id, title, type, passing_score, max_attempts, duration_minutes, negative_marking, is_active]
    QuizInput:
      type: object
      properties:
        topic_id:
          type: ["string", "null"]
        chapter_id:
          type: ["string", "null"]
        subject_id:
          type: ["string", "null"]
        title:
          type: string
        type:
          type: string
          enum: [lecture_quiz, dpp, pyqs, topic_test, chapter_test, subject_test, grand_test]
        passing_score:
          type: number
        max_attempts:
          type: integer
        duration_minutes:
          type: integer
        negative_marking:
          type: number
      required: [title, type, passing_score, max_attempts, duration_minutes, negative_marking]
    QuestionInput:
      type: object
      properties:
        quiz_id:
          type: string
        question_text:
          type: string
        options:
          type: object
          additionalProperties:
            type: string
        correct_answer:
          type: string
        explanation:
          type: ["string", "null"]
        difficulty:
          type: integer
        order_index:
          type: integer
        video_solution_url:
          type: ["string", "null"]
      required: [quiz_id, question_text, options, correct_answer, difficulty, order_index]
    BulkQuestionsInput:
      type: object
      properties:
        quiz_id:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
      required: [quiz_id, questions]
    BulkImportResult:
      type: object
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [imported, failed, errors]
    QrInput:
      type: object
      properties:
        youtube_url:
          type: string
        level:
          type: string
          enum: [subject, chapter, topic, question]
        reference_id:
          type: string
      required: [youtube_url, level, reference_id]
    QrResult:
      type: object
      properties:
        qr_code_url:
          type: string
        youtube_url:
          type: string
      required: [qr_code_url, youtube_url]

    # ── Notes ─────────────────────────────────────────────────────────────
...
    ResetProgressInput:
      type: object
      properties:
        scope:
          type: string
          enum: [topic, chapter, subject, all]
        reference_id:
          type: ["string", "null"]
      required: [scope]
...