# lib/api-zod/src/generated/types/qrResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EdTech Study Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QrResult {
  qr_code_url: string;
  youtube_url: string;
}